@@ -1,5 +1,16 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import App from "./App";
+import { Provider } from "react-redux";
+import { store } from "./app/store";
+import { createRoot } from "react-dom/client";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const container = document.getElementById("root");
+const root = createRoot(container);
+
+root.render(
+  <React.StrictMode>
+    <Provider store={store}>
+      <App />
+    </Provider>
+  </React.StrictMode>
+);
