@@ -7,21 +7,24 @@ import NotFoundPage from "./pages/NotFoundPage";
 import BookDetailPage from "./pages/BookDetailPage";
 import PublicLayout from "./layout/PublicLayout";
 import MThemeProvider from "./theme/MThemeProvider";
-
+import { Provider } from "react-redux";
+import { store } from "./app/store";
 function App() {
   return (
-    <BrowserRouter>
-      <MThemeProvider>
-        <Routes>
-          <Route path="/" element={<PublicLayout />}>
-            <Route index element={<HomePage />} />
-            <Route path="books/:id" element={<BookDetailPage />} />
-            <Route path="reading" element={<ReadingPage />} />
-            <Route path="*" element={<NotFoundPage />} />
-          </Route>
-        </Routes>
-      </MThemeProvider>
-    </BrowserRouter>
+    <Provider store={store}>
+      <BrowserRouter>
+        <MThemeProvider>
+          <Routes>
+            <Route path="/" element={<PublicLayout />}>
+              <Route index element={<HomePage />} />
+              <Route path="books/:id" element={<BookDetailPage />} />
+              <Route path="reading" element={<ReadingPage />} />
+              <Route path="*" element={<NotFoundPage />} />
+            </Route>
+          </Routes>
+        </MThemeProvider>
+      </BrowserRouter>
+    </Provider>
   );
 }
 
