@@ -6,63 +6,58 @@ import SearchForm from "../components/SearchForm";
 import api from "../apiService";
 import { FormProvider } from "../form";
 import { useForm } from "react-hook-form";
-import { Container, Alert, Box, Card, Stack, CardMedia, CardActionArea, Typography, CardContent } from "@mui/material";
-
-
+import {
+  Container,
+  Alert,
+  Box,
+  Card,
+  Stack,
+  CardMedia,
+  CardActionArea,
+  Typography,
+  CardContent,
+} from "@mui/material";
+import { getBooks } from "../features/addList/fetchDataSlice";
+import { useDispatch, useSelector } from "react-redux";
 
 const BACKEND_API = process.env.REACT_APP_BACKEND_API;
 
 const HomePage = () => {
-  const [books, setBooks] = useState([]);
-  const [pageNum, setPageNum] = useState(1);
-  const totalPage = 10;
-  const limit = 10;
+  const dispatch = useDispatch();
+  const { limit, isLoading, books, totalPage, page, pageNum } = useSelector(
+    (store) => store.fetchData
+  );
+  //
 
-  const [loading, setLoading] = useState(false);
-  const [query, setQuery] = useState("");
-  const [errorMessage, setErrorMessage] = useState("");
-
-  const navigate = useNavigate()
+  //
+  const navigate = useNavigate();
   const handleClickBook = (bookId) => {
     navigate(`/books/${bookId}`);
   };
-
-
-
+  //
 
   useEffect(() => {
-    const fetchData = async () => {
-      setLoading(true);
-      try {
-        let url = `/books?_page=${pageNum}&_limit=${limit}`;
-        if (query) url += `&q=${query}`;
-        const res = await api.get(url);
-        setBooks(res.data);
-        setErrorMessage("");
-      } catch (error) {
-        setErrorMessage(error.message);
-      }
-      setLoading(false);
-    };
-    fetchData();
-  }, [pageNum, limit, query]);
+    dispatch(getBooks());
+  }, []);
+
   //--------------form
   const defaultValues = {
-    searchQuery: ""
+    searchQuery: "",
   };
   const methods = useForm({
     defaultValues,
   });
-  const { handleSubmit } = methods;
-  const onSubmit = (data) => {
-    setQuery(data.searchQuery);
-  };
+  // const { handleSubmit } = methods;
+  // // const onSubmit = (data) => {
+  // //   setQuery(data.searchQuery);
+  // // };
   return (
     <Container>
       <Stack sx={{ display: "flex", alignItems: "center", m: "2rem" }}>
-        <Typography variant="h3" sx={{ textAlign: "center" }}>Book Store</Typography>
-        {errorMessage && <Alert severity="danger">{errorMessage}</Alert>}
-        <FormProvider methods={methods} onSubmit={handleSubmit(onSubmit)}>
+        <Typography variant="h3" sx={{ textAlign: "center" }}>
+          Book Store
+        </Typography>
+        <FormProvider methods={methods}>
           <Stack
             spacing={2}
             direction={{ xs: "column", sm: "row" }}
@@ -73,27 +68,30 @@ const HomePage = () => {
             <SearchForm />
           </Stack>
         </FormProvider>
-        <PaginationBar
-          pageNum={pageNum}
-          setPageNum={setPageNum}
-          totalPageNum={totalPage}
-        />
+        <PaginationBar />
       </Stack>
       <div>
-        {loading ? (
-          <Box sx={{ textAlign: "center", color: "primary.main" }} >
+        {isLoading ? (
+          <Box sx={{ textAlign: "center", color: "primary.main" }}>
             <ClipLoader color="inherit" size={150} loading={true} />
           </Box>
         ) : (
-          <Stack direction="row" spacing={2} justifyContent="space-around" flexWrap="wrap">
+          <Stack
+            direction="row"
+            spacing={2}
+            justifyContent="space-around"
+            flexWrap="wrap"
+          >
             {books.map((book) => (
               <Card
-                key={book.id} onClick={() => handleClickBook(book.id)}
+                key={book.id}
+                onClick={() => handleClickBook(book.id)}
                 sx={{
                   width: "12rem",
                   height: "27rem",
                   marginBottom: "2rem",
-                }}>
+                }}
+              >
                 <CardActionArea>
                   <CardMedia
                     component="img"
@@ -104,7 +102,6 @@ const HomePage = () => {
                     <Typography gutterBottom variant="h5" component="div">
                       {`${book.title}`}
                     </Typography>
-
                   </CardContent>
                 </CardActionArea>
               </Card>
